refactor: extract shared internal/fsutil for reflink/device (#49)
internal/link and internal/store each carried byte-identical copies of the
reflink and device-id platform files (~170 LOC across 10 files). Extract them
into internal/fsutil (Reflink, DeviceID) so every future fix lands once.

No behavior change (the FICLONE bug is preserved here and fixed in the next
commit).
